requirements.txt
discord.py
python-dotenv~=1.1.1
dice~=4.0.0